pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = "A cloud-native URL shortener"
authors = ["Koh Jun Jie <koh.jun.jie@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.110.0"
snowflake-id = "^1.0.1"
uvicorn = "^0.27.1"
pybase62 = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
